:art: [need-wallet] moves svg assets to common/icon/*
