+++ src/components/common/icon/coinbase.js
