+++ src/components/common/icon/index.js
@@ -2,10 +2,16 @@ import Login from './login'
 import HomeArt from './home-art'
 import StoreArtTag from './store-art-tag'
 import StoreArtBackground from './store-art-background'
+import Coinbase from './coinbase'
+import Status from './status'
+import Trust from './trust'
 
 export default {
+  Coinbase,
   Login,
   HomeArt,
+  Status,
   StoreArtTag,
   StoreArtBackground,
+  Trust,
 }
