+++ src/components/common/icon/trust.js
